% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_utilities.R
\name{nans.remove}
\alias{nans.remove}
\title{Remove nans from an object}
\usage{
nans.remove(x)
}
\arguments{
\item{x}{a vector, list or data.frame}
}
\description{
this function removes from a given object
}
