% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_utilities.R
\name{relative}
\alias{relative}
\title{Relative abundance calculation}
\usage{
relative(df)
}
\arguments{
\item{df}{a dataframe or a vector of numeric values.}
}
\description{
this function returns the relative proportion of each element based on a vector or dataframe.
}
\details{
This function was designed to perform faster than df/colSums(df) to be used in a bootstrap contest into Renkonen function.
}
\examples{
df <- simulate_data(5,10)
head(df)
df.rel <- relative(df)
head(df.rel)
}
\author{
Wilson Frantine-Silva <wilsonfratine@gmail.com>
}
