% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renkonen.R
\name{renkonen.c}
\alias{renkonen.c}
\title{A caller for calculating the Renkonen's index}
\usage{
renkonen.c(df,as.matrix=T)
}
\arguments{
\item{df}{a dataframe which each column is one site and each row is the absolute abundance of a species}

\item{as.matrix}{a boolean whether it showd be returned as a matrix as well}
}
\description{
This function returns a list of matrix and lists of pairwise comparisons of several sampling sites and its respective Renkonen's similarities.
}
\details{
As renkonen.index is only able to handle vectors, this function parse data.frames and calls the renkonen.index function. This is the basic version for beeing called by a bootstrap handler.
This function was not designed to be used directly at the first place, but users are free to tweeck this.
}
