% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renkonen.R
\name{renkonen.index}
\alias{renkonen.index}
\title{Renkonen equation}
\usage{
renkonen.index(a, b)
}
\arguments{
\item{a}{a vector representing coutings of elements in given group}

\item{b}{same as "a", but from a group to be compared}
}
\description{
the very core of the Renkonen index function. See more in concept and details.
}
\details{
this function calls the relative abundance to avoid users to use absolute countings as input. This function is called by other functions as reikonen() and reikonen.c().
}
\author{
Wilson Frantine-Silva <wilsonfrantine@gmail.com>
}
\concept{Renkonen index is given by the summation of the paralella minimum of to vectors of the same size. In ecological context, it gives the summation of the minimal relative abundances of a species set from two sites. So, if one has sp1 and sp2, representing respectivelly 0.3 and 0.7 at site1 and 0.7 and 0.3 at site2, the Renkonen's index will be S = 0.3 + 0.3 = 0.6. Then, the similarity between site1 and site2 will be 0.6}
