% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renkonen.R
\name{renkonen}
\alias{renkonen}
\title{Renkonen's Similarity index with bootstrap}
\usage{
renkonen(df, boot)
}
\arguments{
\item{df}{a data frame with sampling sites as columns, species as rows and abundances as values}

\item{boot}{an integer with bootstrap replications desired. The default is 0, but something around 1000 seems to be suitable in most cases.}
}
\description{
this function returns both single or bootstrap Renkonen's index for a data frame of counting input. It uses a internal bootstrap function to calculate N matrix. Users might then extract confidence intervals or standard deviation.
}
\details{
This function is the main envelope of Renkonen's index calculations to the users interact with. implements a internal bootstrap replication loop designed to be as fast as possible in a high level language. Remember, bootstrap process are time consuming and the total time will depend on your data set.
}
\examples{
df <- simulate_data(5,10)
S <- renkonen(df)
head(S)
}
\author{
Wilson Frantine-Silva <wilsonfrantine@gmail.com>
}
