% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_utilities.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate a community data}
\usage{
simulate_data(n.sites, n.species, min.ab, mean.ab, seed)
}
\arguments{
\item{n.sites}{the number of sites required or n of columns required}

\item{n.species}{the n of rows representing species}

\item{min.ab}{the expected minimum abundance}

\item{mean.ab}{mean abundance for a given column}

\item{seed}{an integer to be used as random seed.}
}
\description{
this function simulates a single community data using a pseudorandom root, uniform and poison distributions
}
\examples{
simulate_data()
simulate_data(n.species=10,n.sites=20)
simulate_data(n.species=10,n.sites=20, 
    min.ab=10, mean.ab=100, seed=123)
}
