% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_utilities.R
\name{tree_consensus}
\alias{tree_consensus}
\title{Compute tree consensus for multiple trees}
\usage{
tree_consensus(trees, type, p)
}
\arguments{
\item{trees}{an object of the class multiPhylo}

\item{type}{a type of consensus. Default is "mcc" (Maximum Credibility Tree), but majority rule ("majority") or "strict" rules are also avaliable. In case of majority rule, the user also might use the param "p".}

\item{p}{the proportion of concordance for some specific node. Default is 0.5, but users might change that between 0 and 1. If 1 is seted the algoritms behaves as strict rule.}
}
\description{
this function returns a consensus tree of the class phylo. I can be used to be saved or plot with diferent functions and pakcages
}
\details{
this function is an wrap of ape and phangorn packages
}
\examples{
df <- simulate_data()
S <- renkonen(df, 100)
trees <- upgma(S)
mcc_tree <- tree_consensus(trees, "mcc")
}
