% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgma.R
\name{upgma.c}
\alias{upgma.c}
\title{UPGMA calc core}
\usage{
upgma.c(df, method = "average", ...)
}
\arguments{
\item{df}{a squar data frame with similarity index as values}

\item{method}{same methods as in hclust function}

\item{...}{ignored}
}
\description{
this function is just a wraper for the hclut function which calculates an upgma dendrogram from a distance matrix. The function also converts a the hclust object in a phylo object.
}
\seealso{
stats::hclust()
}
\keyword{internal}
