% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgma.R
\name{upgma}
\alias{upgma}
\title{UPGMA calculation}
\usage{
upgma(x, method = "average", ...)
}
\arguments{
\item{x}{a distance matrix or a list of distance matrix as outputed by the renkonen function}

\item{method}{the UPGMA summation method. It uses the same as hclust function}

\item{...}{ignored}
}
\description{
it takes a tree or a list of trees and returns a phylo object (from ape package) with x trees as inputed.
}
\seealso{
renkonen(), stats::hclust()
}
\concept{the function is based on stats:hclust.}
